guidance-for-predictive-maintenance/pyproject.toml
[tool.poetry]
name = "mmt-predictive-maintenance"
version = "1.0.0"
description = "MMT Predictive Maintenance Solution"
authors = ["Amazon Web Services"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
jinja2 = "*"
markdown-to-json = "*"
mypy = "*"
mypy-extensions = "*"
pre-commit = "*"
pycln = "*"
pylint = "*"
types-pyyaml = "*"
wheel = "*"
attrs = "*"
aws_lambda_powertools = { extras = ["tracer", "validation"], version = ">=3.3.0" }
aws-cdk-lib = "^2.204.0"
boto3 = "^1.39.3"
botocore = "^1.39.3"
boto3-stubs = {extras = ["essential", "cloudformation", "dynamodb", "s3", "sagemaker", "secretsmanager", "ssm", "stepfunctions"], version = "*"}
constructs = "^10.4.2"
typing-inspect = "*"
typing-extensions = "*"
cdk-monitoring-constructs = "^9.12.0"
pandas-stubs = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["attrs_plugin"]
